+++ contracts/src/models/allowed_app.cairo
@@ -1,10 +1,10 @@
 use starknet::ContractAddress;
 
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct AllowedApp {
     #[key]
-    game_id: usize,
+    game_id: u32,
     #[key]
     contract: ContractAddress,
     is_allowed: bool
