+++ contracts/src/models/allowed_color.cairo
@@ -1,9 +1,9 @@
 // if the color is allowed
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct AllowedColor {
     #[key]
-    game_id: usize,
+    game_id: u32,
     #[key]
     color: u32,
     is_allowed: bool
@@ -11,10 +11,10 @@ struct AllowedColor {
 
 // the color found in what index
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct PaletteColors {
     #[key]
-    game_id: usize,
+    game_id: u32,
     #[key]
     idx: u32,
     color: u32
@@ -22,10 +22,10 @@ struct PaletteColors {
 
 // if the color is already in the palette
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct InPalette {
     #[key]
-    game_id: usize,
+    game_id: u32,
     #[key]
     color: u32,
     value: bool
@@ -33,9 +33,9 @@ struct InPalette {
 
 // number of colors in the game's palette
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct GamePalette {
     #[key]
-    game_id: usize,
-    length: usize
+    game_id: u32,
+    length: u32
 }
