+++ contracts/src/models/board.cairo
@@ -2,17 +2,17 @@ use pixelaw::core::utils::Position;
 use starknet::{ContractAddress};
 
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct Board {
     #[key]
-    id: usize,
+    id: u32,
     origin: Position,
     width: u32,
     height: u32,
 }
 
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct PWarPixel {
     #[key]
     position: Position,
@@ -21,11 +21,11 @@ struct PWarPixel {
 
 
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 struct GameId {
     #[key]
     x: u32,
     #[key]
     y: u32,
-    value: usize
+    value: u32
 }
